redditcollector/requirements.txt
pandas>=1
pandas_gbq
psaw
pyyaml